ai_module/requirements.txt
 fastapi
uvicorn[standard]
python-multipart
aiofiles
requests
# ML libs (may be large)
torch
torchvision
torchgeo
pillow
